requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.5
chardet==4.0.0
discord.py==1.7.1
ffmpeg-python==0.2.0
future==0.18.2
idna==3.1
multidict==5.1.0
playsound==1.2.2
pycparser==2.20
pydub==0.25.1
PyNaCl==1.4.0
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
yt-dlp
